# Conflicts:
#	composer.json
#	composer.lock

+++ src/Migration/Migration1745505195RemoveEnableSearchImpressionsConfig.php
@@ -14,8 +14,15 @@ class Migration1745505195RemoveEnableSearchImpressionsConfig extends MigrationSt
         return 1745505195;
     }
 
+    /**
+     * @throws \Doctrine\DBAL\Exception
+     */
     public function update(Connection $connection): void
     {
+        if (!$connection->createSchemaManager()->tableExists('nosto_integration_config')) {
+            return;
+        }
+
         $sql = <<<SQL
             DELETE FROM `nosto_integration_config`
             WHERE `configuration_key` = 'enableSearchImpressions'
